#255 - refactoring
